+++ selector/types.go
@@ -48,8 +48,8 @@ type monitorCandidate struct {
 	HasMetrics               bool
 	IsHealthy                bool
 	RTT                      float64
-	Priority                 float64 // Database-calculated monitor priority (RTT + step accuracy)
-	Count                    int64   // Number of data points from GetMonitorPriority query
+	Priority                 int   // Database-calculated monitor priority (RTT + step accuracy)
+	Count                    int64 // Number of data points from GetMonitorPriority query
 	ConstraintViolationType  *string
 	ConstraintViolationSince *time.Time
 }
